docs: update AGENTS.md and agent documentation for v0.1.31
Refresh agent guidelines and documentation to reflect current
project state and workflow improvements.

Changes:
- Update AGENTS.md with skills compaction results
- Add agent_docs/LESSONS.md for session learnings
- Refresh README.md, CONTRIBUTING.md, SECURITY.md, DEPLOYMENT.md
- Update agent_docs/running_tests.md with nextest patterns
- Update agent_docs/git_workflow.md with branch protection notes
