+++ memory-core/src/context/accumulator.rs
@@ -418,193 +418,3 @@ impl Default for BundleAccumulator {
         Self::default_config()
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::episode::Episode;
-    use crate::types::{TaskContext, TaskType};
-
-    fn create_test_episode(description: &str, days_ago: i64) -> std::sync::Arc<Episode> {
-        let mut episode = Episode::new(
-            description.to_string(),
-            TaskContext::default(),
-            TaskType::Debugging,
-        );
-        episode.start_time = Utc::now() - chrono::Duration::days(days_ago);
-        std::sync::Arc::new(episode)
-    }
-
-    #[test]
-    fn test_new_accumulator() {
-        let acc = BundleAccumulator::default_config();
-        assert!(acc.is_empty());
-        assert_eq!(acc.size(), 0);
-        assert!(!acc.is_full());
-    }
-
-    #[test]
-    fn test_add_single_item() {
-        let mut acc = BundleAccumulator::default_config();
-        let episode = create_test_episode("test", 0);
-        let item = ContextItem::from_episode(episode, 0.8);
-
-        let result = acc.add(item);
-        assert!(result.accepted);
-        assert_eq!(acc.size(), 1);
-    }
-
-    #[test]
-    fn test_reject_below_threshold() {
-        let config = BundleConfig {
-            min_salience_threshold: 0.5,
-            ..BundleConfig::default()
-        };
-        let mut acc = BundleAccumulator::new(config);
-
-        let episode = create_test_episode("test", 0);
-        let item = ContextItem::from_episode(episode, 0.3); // Below threshold
-
-        let result = acc.add(item);
-        assert!(!result.accepted);
-        assert_eq!(acc.stats().total_rejected, 1);
-        assert!(result.rejection_reason.is_some());
-    }
-
-    #[test]
-    fn test_eviction_when_full() {
-        let config = BundleConfig {
-            max_items: 2,
-            ..BundleConfig::default()
-        };
-        let mut acc = BundleAccumulator::new(config);
-
-        // Add two items
-        let ep1 = create_test_episode("high priority", 0);
-        let item1 = ContextItem::from_episode(ep1, 0.9);
-        acc.add(item1);
-
-        let ep2 = create_test_episode("medium priority", 10);
-        let item2 = ContextItem::from_episode(ep2, 0.5);
-        acc.add(item2);
-
-        assert_eq!(acc.size(), 2);
-        assert!(acc.is_full());
-
-        // Add third item - should evict lowest priority
-        let ep3 = create_test_episode("new high", 0);
-        let item3 = ContextItem::from_episode(ep3, 0.95);
-        let result = acc.add(item3);
-
-        assert!(result.accepted);
-        assert!(result.evicted_id.is_some());
-        assert_eq!(acc.stats().total_evicted, 1);
-    }
-
-    #[test]
-    fn test_to_bundle_sorted() {
-        let mut acc = BundleAccumulator::default_config();
-
-        // Add items with different priorities
-        let ep1 = create_test_episode("low", 30);
-        let item1 = ContextItem::from_episode(ep1, 0.3);
-        acc.add(item1);
-
-        let ep2 = create_test_episode("high", 0);
-        let item2 = ContextItem::from_episode(ep2, 0.9);
-        acc.add(item2);
-
-        let ep3 = create_test_episode("medium", 15);
-        let item3 = ContextItem::from_episode(ep3, 0.5);
-        acc.add(item3);
-
-        let bundle = acc.to_bundle();
-
-        // Should be sorted by priority (highest first)
-        assert!(bundle[0].priority() >= bundle[1].priority());
-        assert!(bundle[1].priority() >= bundle[2].priority());
-    }
-
-    #[test]
-    fn test_clear() {
-        let mut acc = BundleAccumulator::default_config();
-        let episode = create_test_episode("test", 0);
-        let item = ContextItem::from_episode(episode, 0.8);
-        acc.add(item);
-
-        acc.clear();
-        assert!(acc.is_empty());
-        assert_eq!(acc.stats().total_added, 0); // Reset stats
-    }
-
-    #[test]
-    fn test_from_episodes() {
-        let episodes = vec![
-            create_test_episode("ep1", 0),
-            create_test_episode("ep2", 10),
-            create_test_episode("ep3", 20),
-        ];
-
-        let bundle = BundleAccumulator::from_episodes(episodes, |_| 0.8);
-
-        assert_eq!(bundle.len(), 3);
-    }
-
-    #[test]
-    fn test_config_validation() {
-        let valid = BundleConfig::default();
-        assert!(valid.validate().is_ok());
-
-        let invalid = BundleConfig {
-            max_items: 0,
-            ..BundleConfig::default()
-        };
-        assert!(invalid.validate().is_err());
-    }
-
-    #[test]
-    fn test_stats_update() {
-        let mut acc = BundleAccumulator::default_config();
-
-        let ep1 = create_test_episode("test", 5);
-        let item1 = ContextItem::from_episode(ep1, 0.7);
-        acc.add(item1);
-
-        let ep2 = create_test_episode("test2", 0);
-        let item2 = ContextItem::from_episode(ep2, 0.8);
-        acc.add(item2);
-
-        let _bundle = acc.to_bundle();
-        let stats = acc.stats();
-
-        assert_eq!(stats.current_size, 2);
-        assert!(stats.average_salience > 0.0);
-        assert!(stats.oldest_timestamp.is_some());
-        assert!(stats.newest_timestamp.is_some());
-    }
-
-    #[test]
-    fn test_contains_and_get() {
-        let mut acc = BundleAccumulator::default_config();
-        let episode = create_test_episode("test", 0);
-        let id = episode.episode_id;
-        let item = ContextItem::from_episode(episode, 0.8);
-        acc.add(item);
-
-        assert!(acc.contains(id));
-        assert!(acc.get(id).is_some());
-    }
-
-    #[test]
-    fn test_remove() {
-        let mut acc = BundleAccumulator::default_config();
-        let episode = create_test_episode("test", 0);
-        let id = episode.episode_id;
-        let item = ContextItem::from_episode(episode, 0.8);
-        acc.add(item);
-
-        assert!(acc.remove(id));
-        assert!(acc.is_empty());
-        assert!(!acc.remove(id)); // Already removed
-    }
-}
